OAK-1620 Index cost calculation integer overflow
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1581892 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/property/PropertyIndexLookup.java
@@ -62,7 +62,7 @@ public class PropertyIndexLookup {
     /**
      * The cost overhead to use the index in number of read operations.
      */
-    private static final int COST_OVERHEAD = 2;
+    private static final double COST_OVERHEAD = 2;
     
     /**
      * The maximum cost when the index can be used.
